Imperial aramaic (#187)
* [arc] Listing the correct scripts for Imperial Aramaic:

1.  The original Aramaic script (`armi`).
2.  The square script as in Biblical Aramaic (`hebr`).
3.  Classical Syriac/Assyrian Neo-Aramaic (`syrc`) descended from (1).

This correctly assigns the entries to their respective lexicons. Most
of pronunciations are available for (2), with very minor number of
entries for (1) and (3).

* [arc] Listing the correct scripts for Imperial Aramaic (continuing the
previous commit which was partial):

1.  The original Aramaic script (`armi`).
2.  The square script as in Biblical Aramaic (`hebr`).
3.  Classical Syriac/Assyrian Neo-Aramaic (`syrc`) descended from (1).

This correctly assigns the entries to their respective lexicons. Most
of pronunciations are available for (2), with very minor number of
entries for (1) and (3).

* Updated CHANGELOG with #186.